don't rely on global $provider anymore
The statsd::params class was relying on a global $provider variable
whereas the statsd class tooks it from it's parameters. This is now also
a class parameter in the params class. statsd isn't inheriting from
statsd::params anymore (was useless before because all parameters were
explicitly set in the statsd class anyway).