Move var determination from params back to init
According to the puppet docs base classes "must have no parameters"
and they are evaluated before the inheriting class
so any parameters are unavailable.

This is why the current code errors on debian/ubuntu
(the $provider var is empty at the time the params class checks it).

So in order to keep those parameters just move the logic
back to the main class for simplicity.